requirements.txt
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-llms-huggingface>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.0
python-dotenv>=1.0.0
nest-asyncio>=1.5.8
datasets>=2.18.0
logging>=0.5.1.2
